chore: refactor test.utils.ts and PacketUtils into exported consts
